Change the Environment to use url::Url
Get rid of most of the CGI PATH_INFO etc remnants
and use a better data structure (url::Url). Document
semantics related to what `None` means in the context
of a url (parsing error) and how RFC2616§5.1.2's "*"
is represented.

Add #[deriving(FromPrimitive)] to the `Method` enum
so it can be translated to more easily.